Update angular routing strategy
