+++ generators/client/templates/angular/src/main/webapp/app/layouts/_index.ts
@@ -21,10 +21,10 @@ export * from './error/error.route';
 export * from './main/main.component';
 export * from './footer/footer.component';
 export * from './navbar/navbar.component';
+export * from './navbar/navbar.route';
 <%_ if (enableTranslation) { _%>
 export * from './navbar/active-menu.directive';
 <%_ } _%>
 export * from './profiles/page-ribbon.component';
 export * from './profiles/profile.service';
 export * from './profiles/profile-info.model';
-export * from './layout-routing.module';
